third_party/WebKit/LayoutTests/media/track/track-load-from-element-readyState.html

<!DOCTYPE html>
<title>Tests the load event on HTMLTrackElement and LOADED readyState on TextTrack when src is set on the element.</title>
<video>
    <track src="captions-webvtt/tc004-webvtt-file.vtt" default>
</video>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
async_test(function(t) {
    var track = document.querySelector("track");
    track.onload = t.step_func_done(function() {
        assert_equals(track.readyState, HTMLTrackElement.LOADED);
    });
});
</script>
